[UPD] Models summed in one file
After removing the redundant response structs, there is no need to split the few models into so many files.
