+++ api/handler/order_item_cancel.go
@@ -9,7 +9,7 @@ import (
 	"github.com/plaenkler/booklooker/api/models"
 )
 
-func PutOrderItemCancel(token string, req *models.OrderItemCancelRequest) (*models.GeneralResponse, error) {
+func PutOrderItemCancel(token string, req *models.OrderItemCancelRequest) (*models.GlobalResponse, error) {
 	url := baseURL + models.OrderItemCancelPath + "?token=" + token
 	jsonReq, err := json.Marshal(req)
 	if err != nil {
@@ -30,7 +30,7 @@ func PutOrderItemCancel(token string, req *models.OrderItemCancelRequest) (*mode
 	if err != nil {
 		return nil, err
 	}
-	var orderItemCancelResp models.GeneralResponse
+	var orderItemCancelResp models.GlobalResponse
 	err = json.Unmarshal(jsonResp, &orderItemCancelResp)
 	if err != nil {
 		return nil, err
