--- api/models/article.go
@@ -1,7 +1,0 @@
-package models
-
-const ArticlePath = "article"
-
-type ArticleRequest struct {
-	OrderNo string `json:"orderNo"`
-}
