--- api/models/article_list.go
@@ -1,10 +1,0 @@
-package models
-
-const ArticleListPath = "article_list"
-
-type ArticleListRequest struct {
-	ReturnType string `json:"field"`
-	ShowPrice  bool   `json:"showPrice"`
-	ShowStock  bool   `json:"showStock"`
-	MediaType  byte   `json:"mediaType,omitempty"`
-}
