--- api/models/article_status.go
@@ -1,7 +1,0 @@
-package models
-
-const ArticleStatusPath = "article_status"
-
-type ArticleStatusRequest struct {
-	OrderNo string `json:"orderNo"`
-}
