--- api/models/file_status.go
@@ -1,7 +1,0 @@
-package models
-
-const FileStatusPath = "file_status"
-
-type FileStatusRequest struct {
-	Filename string `json:"filename"`
-}
