--- api/models/import_status.go
@@ -1,3 +1,0 @@
-package models
-
-const ImportStatusPath = "import_status"
