--- api/models/order_cancel.go
@@ -1,7 +1,0 @@
-package models
-
-const OrderCancelPath = "order_cancel"
-
-type OrderCancelRequest struct {
-	OrderID string `json:"orderId"`
-}
