--- api/models/order_item_cancel.go
@@ -1,8 +1,0 @@
-package models
-
-const OrderItemCancelPath = "order_item_cancel"
-
-type OrderItemCancelRequest struct {
-	OrderItemID string `json:"orderItemId"`
-	MediaType   byte   `json:"mediaType"`
-}
