--- api/models/order_status.go
@@ -1,8 +1,0 @@
-package models
-
-const OrderStatusPath = "order_status"
-
-type OrderStatusRequest struct {
-	OrderID string `json:"orderId"`
-	Status  string `json:"status"`
-}
